+++ partiql-planner/src/testFixtures/resources/inputs/basics/case.sql
@@ -85,8 +85,7 @@ CASE t_item.t_string
 END;
 
 --#[case-when-11]
--- type: (int|null|missing)
--- TODO should really be (int|missing) but our translation of coalesce doesn't consider types.
+-- type: (int|missing)
 COALESCE(CAST(t_item.t_string AS INT), 1);
 
 -- -----------------------------
@@ -255,6 +254,42 @@ CASE t_item.t_string
     WHEN 'a' THEN MISSING
     WHEN 'b' THEN MISSING
     ELSE MISSING
+END;
+
+-- -----------------------------
+--  Any Branches
+-- -----------------------------
+
+--#[case-when-31]
+-- type: (any)
+CASE t_item.t_string
+    WHEN 'a' THEN t_item.t_any
+    WHEN 'b' THEN t_item.t_int32
+    ELSE NULL
+END;
+
+--#[case-when-32]
+-- type: (any)
+CASE t_item.t_string
+    WHEN 'a' THEN t_item.t_int32
+    WHEN 'b' THEN t_item.t_any
+    ELSE NULL
+END;
+
+--#[case-when-33]
+-- type: (any)
+CASE t_item.t_string
+    WHEN 'a' THEN t_item.t_int32
+    WHEN 'b' THEN NULL
+    ELSE t_item.t_any
+END;
+
+--#[case-when-34]
+-- type: (any)
+CASE t_item.t_string
+    WHEN 'a' THEN t_item.t_int32_null
+    WHEN 'b' THEN t_item.t_any
+    ELSE t_item.t_any
 END;
 
 -- -----------------------------